+++ packages/components/src/components/telekom/app-shell/app-shell.tsx
@@ -33,6 +33,7 @@ export class Shell {
   @Prop() addonNavigation?: any = [];
   @Prop() activeRouteId?: string = '';
   @Prop() activeSectorId?: string = '';
+  @Prop() userMenuAriaLabel?: string;
   @Prop() sticky?: boolean = false;
   @State() scrolled: boolean = false;
   /** (optional) Injected CSS styles */
@@ -68,6 +69,7 @@ export class Shell {
               activeSectorId={this.activeSectorId}
               claimLang={this.claimLang}
               sticky={this.sticky}
+              userMenuAriaLabel={this.userMenuAriaLabel}
             ></scale-app-header>
           )}
           <main class="content">