+++ packages/storybook-vue/stories/components/beta-brand-header-navigation/SlotMainNavWithFlyout.vue
@@ -52,18 +52,6 @@
               <li><a href="https://example.com">Third Level</a></li>
             </ul>
           </scale-telekom-mega-menu-column>
-          <scale-telekom-mega-menu-column>
-            <scale-icon-home-home slot="icon"></scale-icon-home-home>
-            <a href="https://example.com" slot="heading">Second Level</a>
-            <ul>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-            </ul>
-          </scale-telekom-mega-menu-column>                       
         </scale-telekom-mega-menu>
       </scale-telekom-nav-flyout>
     </scale-telekom-nav-item>
@@ -107,18 +95,6 @@
               <li><a href="https://example.com">Third Level</a></li>
             </ul>
           </scale-telekom-mega-menu-column>
-          <scale-telekom-mega-menu-column>
-            <scale-icon-home-home slot="icon"></scale-icon-home-home>
-            <a href="https://example.com" slot="heading">Second Level</a>
-            <ul>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-              <li><a href="https://example.com">Third Level</a></li>
-            </ul>
-          </scale-telekom-mega-menu-column>                      
         </scale-telekom-mega-menu>
       </scale-telekom-nav-flyout>
     </scale-telekom-nav-item>