+++ packages/storybook-vue/stories/components/beta-brand-header-navigation/fixtures.js
@@ -360,6 +360,5 @@ export const iconNavigation = [
 ];
 
 export const sectorNavigation = [
-  { name: 'Personal', id: 'Personal38', href: '#personal' },
   { name: 'Business', id: 'Business39', href: '#business' },
 ];